+++ yash-env/CHANGELOG.md
@@ -14,6 +14,8 @@ A _private dependency_ is used internally and not visible to downstream users.
 ### Added
 
 - `function::FunctionBody`: Trait for the body of a shell function.
+- `function::FunctionBodyObject`: Dyn-compatible adapter for the `FunctionBody`
+  trait.
 - `parser::IsName`: Wrapper for a function that checks if a string is a valid
   variable name. This allows modules to check variable names without directly
   depending on the `yash-syntax` crate.